+++ platform/scms/app/routes/_auth/utils.tsx
@@ -1,4 +1,4 @@
-import { firebase, github, google, okta, orcid } from '@curvenote/scms-core';
+import { firebase, github, google, okta, orcid, bluesky } from '@curvenote/scms-core';
 
 export function getProviderUI({
   provider,
@@ -25,6 +25,10 @@ export function getProviderUI({
     case 'orcid':
       return (
         <orcid.LoginUI disabled={submitting} setSubmitting={setSubmitting} className="w-full" />
+      );
+    case 'bluesky':
+      return (
+        <bluesky.LoginUI disabled={submitting} setSubmitting={setSubmitting} className="w-full" />
       );
     case 'firebase':
       return (