+++ packages/scms-core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @curvenote/scms-core
 
+## 0.15.5
+
+### Patch Changes
+
+- [#831](https://github.com/curvenote/curvenote/pull/831) [`010d781`](https://github.com/curvenote/curvenote/commit/010d7818a46e7265b47ca1959154901ccc79549c) Thanks [@stevejpurves](https://github.com/stevejpurves)! - Moving schemas to central location
+
+- [#814](https://github.com/curvenote/curvenote/pull/814) [`a350156`](https://github.com/curvenote/curvenote/commit/a35015615fa37b752938ea93e02a066584740414) Thanks [@stevejpurves](https://github.com/stevejpurves)! - Adding BlueSky auth provider
+
+- [#803](https://github.com/curvenote/curvenote/pull/803) [`3025543`](https://github.com/curvenote/curvenote/commit/302554357f2233caad98fd9d28dfe7cad82397e1) Thanks [@stevejpurves](https://github.com/stevejpurves)! - Improving links in all slack messages
+
+- Updated dependencies []:
+  - @curvenote/scms-db@0.15.5
+
 ## 0.15.4
 
 ### Patch Changes